Function ink_lang::codegen::execute_message

pub fn execute_message<Storage, Output, F>(
    config: ExecuteMessageConfig, 
    f: F
) -> Result<(), DispatchError> where
    Storage: SpreadLayout + ContractEnv,
    Output: Encode + 'static,
    F: FnOnce(&mut Storage) -> Output, 

Executes the given &mut self message closure.

Note

The closure is supposed to already contain all the arguments that the real message requires and forwards them.