Function ink_lang::codegen::finalize_message

pub fn finalize_message<Contract, R>(
    success: bool, 
    contract: &Contract, 
    config: ExecuteMessageConfig, 
    result: &R
) -> Result<(), DispatchError> where
    Contract: SpreadLayout,
    R: Encode + 'static, 

Finalizes an ink! message call with the given configuration.

This dispatches into fallible and infallible message finalization depending on the given success state.

If the message call was successful the return value is simply returned and cached storage is pushed back to the contract storage.

If the message call failed the return value result is returned instead and the transaction is signalled to be reverted.

Note

This work around that splits executing an ink! message into initiate and finalize phases was needed due to the fact that is_result_type and is_result_err macros do not work in generic contexts.