Function ink_lang::codegen::initiate_message

pub fn initiate_message<Contract>(
    config: ExecuteMessageConfig
) -> Result<Contract, DispatchError> where
    Contract: SpreadLayout + ContractEnv, 

Initiates an ink! message call with the given configuration.

Returns the contract state pulled from the root storage region upon success.

Note

This work around that splits executing an ink! message into initiate and finalize phases was needed due to the fact that is_result_type and is_result_err macros do not work in generic contexts.