Trait ink_lang::codegen::EventLenTopics

pub trait EventLenTopics {
    type LenTopics;
}

Stores the number of event topics of the ink! event definition.

Associated Types

type LenTopics

Type denoting the number of event topics.

Note

We use an associated type instead of an associated constant here because Rust does not yet allow for generics in constant parameter position which would be required in the EventRespectsTopicLimit trait definition. As soon as this is possible in Rust we might change this to a constant.