Function ink_lang::execute_constructor

pub fn execute_constructor<C, F>(
    enables_dynamic_storage_allocator: EnablesDynamicStorageAllocator, 
    f: F
) -> Result<(), DispatchError> where
    C: Constructor,
    F: FnOnce() -> <C as FnState>::State, 

Executes the given constructor closure.

Note

The closure is supposed to already contain all the arguments that the real constructor message requires and forwards them.