Struct ink_lang::Contract

pub struct Contract<Constrs, Msgs, Phase> { /* fields omitted */ }

The contract definition.

Dispatch

Constructor Dispatch

Retrieve function selector s.

Find selected constructor C given s:

If C was found:

Decode input arguments i according to C's requirements

Initialize dynamic storage allocator for instantiation

Call C(i), returns storage instance S

Push state of S to contract storage

Finalize dynamic storage allocator

Exit

Otherwise:

Panic that the provided constructor selector was invalid

Message Dispatch

Retrieve function selector s.

Find selected message M given s:

If M was found:

Decode input arguments i according to M's requirements

Initialize dynamic storage allocator for calls

Pull storage instance S from contract storage

Call M(s, i), returns return value R

If M has mutable access to the contract storage:

push mutated state of S back to contract storage

Finalize dynamic storage allocator

If R is not of type ():

return R

Exit

Otherwise:

Panic that the provided message selector was invalid

Implementations

impl Contract<(), (), EmptyPhase>

pub fn build() -> Contract<EmptyDispatchList, EmptyDispatchList, BuildPhase>

Creates a new contract builder.

impl<Constrs> Contract<Constrs, EmptyDispatchList, BuildPhase>

pub fn register_constructor<C>(
self
) -> Contract<DispatchList<MsgCon<C>, Constrs>, EmptyDispatchList, BuildPhase> where
C: Constructor,

Registers a new constructor for the contract.

impl<Constrs, Messages> Contract<Constrs, Messages, BuildPhase>

pub fn register_message<M>(
self
) -> Contract<Constrs, DispatchList<MsgRef<M>, Messages>, BuildPhase> where
M: MessageRef,

Registers a new &self message for the contract.

pub fn register_message_mut<M>(
self
) -> Contract<Constrs, DispatchList<MsgMut<M>, Messages>, BuildPhase> where
M: MessageMut,

Registers a new &mut self message for the contract.

pub fn finalize(self) -> Contract<Constrs, Messages, FinalPhase>

Finalizes the construction of the contract.

impl<Constrs, Messages> Contract<Constrs, Messages, FinalPhase> where
Constrs: Dispatch,

pub fn on_instantiate(self, call_data: &CallData) -> Result<(), DispatchError>

Instantiates the contract.

impl<Constrs, Messages> Contract<Constrs, Messages, FinalPhase> where
Messages: Dispatch,

pub fn on_call(self, call_data: &CallData) -> Result<(), DispatchError>

Calls a contract message.