Trait ink_lang::EmitEvent

pub trait EmitEvent<C> where
    C: BaseEvent, {
    fn emit_event<E>(self, event: E)
    where
        E: Into<<C as BaseEvent>::Type>;
}

Implemented by contracts in order to override env().emit_event(..) syntax for emitting of ink! contract events.

Dev Note

Normally we’d try to define traits like these in the compagnion ink_lang crate, however, due to Rust’s orphan rules we must define this trait here.

Required methods

fn emit_event<E>(self, event: E) where
E: Into<<C as BaseEvent>::Type>,

Emits an event that can be trivially converted into the base event.

Loading content...