Trait ink_lang::Env

pub trait Env {
    type EnvAccess;
    fn env(self) -> Self::EnvAccess;
}

Simplifies interaction with the host environment via self.

Note

This is generally implemented for storage structs that include their environment in order to allow the different dispatch functions to use it for returning the contract’s output.

Associated Types

type EnvAccess

The access wrapper.

Required methods

fn env(self) -> Self::EnvAccess

Accesses the environment with predefined environmental types.