Trait ink_lang::StaticEnv

pub trait StaticEnv {
    type EnvAccess;
    fn env() -> Self::EnvAccess;
}

Simplifies interaction with the host environment via Self.

Note

This is generally implemented for storage structs that include their environment in order to allow the different dispatch functions to use it for returning the contract’s output.

Associated Types

type EnvAccess

The access wrapper.

Loading content...
Required methods

fn env() -> Self::EnvAccess

Accesses the environment with predefined environmental types.

Loading content...