Crate ink_lang

Structs

Contract
The contract definition.

ContractBuilder
Simplifies declaration of a smart contract.

DispatchList
A list of dispatchers.

DispatchRetCode
A return code indicating success or error in a compact form.

Dispatcher
Dispatcher for storage preserving messages.

DispatcherMut
Dispatcher for potentially storage mutating messages and constructors.

EnvAccess
A typed accessor to the environment.

UnreachableDispatcher
A dispatcher that shall never dispatch.

Enums

DispatchError
A dispatch error.

DispatchMode
The contract dispatch mode.

Traits

Dispatch
Types implementing this trait can handle contract calls.

DispatchUsingMode
Trait implemented by contracts themselves in order to provide a clean interface for the C-ABI specified call and create functions to forward calls to.

Env
Simplifies interaction with the host environment via self.

FnInput
Dispatchable functions that have inputs.

FnOutput
Dispatchable functions that have an output.

FnSelector
The selector of dispatchable functions.

ForwardCall
Implemented by contracts that are compiled as dependencies.

ForwardCallMut
Implemented by contracts that are compiled as dependencies.

GenerateAbi
Types implementing this trait can generate their metadata.

InstantiateTestable
Trait implemented by contracts to make them testable.

Message
Types implementing this are messages that may only read from storage.

PushDispatcher
Types able to push another dispatcher to themselves.

StaticEnv
Simplifies interaction with the host environment via Self.

Storage
Types implementing this trait are storage structs.

ToAccountId
Implemented by contracts that are compiled as dependencies.

Type Definitions

DispatchResult
A dispatch result.

DispatchableFn
A function with the signature able to handle storage preserving calls of messages or constructors.

DispatchableFnMut
A function with the signature able to handle potentially storage mutating calls of messages or constructors.

EmptyDispatchList
A simple type definition to view the single unreachable dispatcher as list.

Attribute Macros

contract