Trait ink_lang::codegen::ContractRootKey

pub trait ContractRootKey {
    const ROOT_KEY: Key;
}

The root key of the ink! smart contract.

Note

This is the key where storage allocation, pushing and pulling is rooted using the SpreadLayout and SpreadAllocate traits primarily.

This trait is automatically implemented by the ink! codegen.

The existence of this trait allows to customize the root key in future versions of ink! if needed.

Associated Constants

const ROOT_KEY: Key