Struct ink_lang::EnvAccess

pub struct EnvAccess<'a, T> { /* fields omitted */ }

A typed accessor to the environment.

This allows ink! messages to make use of the environment efficiently and user friendly while also maintaining access invariants.

Implementations

impl<'a, T> EnvAccess<'a, T> where
T: Environment,
<T as Environment>::ChainExtension: ChainExtensionInstance,

pub fn extension(
self
) -> <<T as Environment>::ChainExtension as ChainExtensionInstance>::Instance

Allows to call one of the available defined chain extension methods.

impl<'a, T> EnvAccess<'a, T> where
T: Environment,

pub fn caller(self) -> T::AccountId

Returns the address of the caller of the executed contract.

Note

For more details visit: ink_env::caller

pub fn transferred_balance(self) -> T::Balance

Returns the transferred balance for the contract execution.

Note

For more details visit: ink_env::transferred_balance

pub fn weight_to_fee(self, gas: u64) -> T::Balance

Returns the price for the specified amount of gas.

Note

For more details visit: ink_env::weight_to_fee

pub fn gas_left(self) -> T::Balance

Returns the amount of gas left for the contract execution.

Note

For more details visit: ink_env::gas_left

pub fn block_timestamp(self) -> T::Timestamp

Returns the timestamp of the current block.

Note

For more details visit: ink_env::block_timestamp

pub fn account_id(self) -> T::AccountId

Returns the account ID of the executed contract.

Note

For more details visit: ink_env::account_id

pub fn balance(self) -> T::Balance

Returns the balance of the executed contract.

Note

For more details visit: ink_env::balance

pub fn rent_allowance(self) -> T::Balance

Returns the current rent allowance for the executed contract.

Note

For more details visit: ink_env::rent_allowance

pub fn block_number(self) -> T::BlockNumber

Returns the current block number.

Note

For more details visit: ink_env::block_number

pub fn minimum_balance(self) -> T::Balance

Returns the minimum balance that is required for creating an account.

Note

For more details visit: ink_env::minimum_balance

pub fn tombstone_deposit(self) -> T::Balance

Returns the tombstone deposit for the contracts chain.

Note

For more details visit: ink_env::tombstone_deposit

pub fn set_rent_allowance(self, new_value: T::Balance)

Sets the rent allowance of the executed contract to the new value.

Note

For more details visit: ink_env::set_rent_allowance

pub fn invoke_contract<Args>(
self,
params: &CallParams<T, Args, ()>
) -> Result<()> where
Args: Encode,

Invokes a contract message.

Note

For more details visit: ink_env::invoke_contract

pub fn eval_contract<Args, R>(
self,
params: &CallParams<T, Args, ReturnType<R>>
) -> Result<R> where
Args: Encode,
R: Decode,

Evaluates a contract message and returns its result.

Note

For more details visit: ink_env::eval_contract

pub fn instantiate_contract<Args, Salt, C>(
self,
params: &CreateParams<T, Args, Salt, C>
) -> Result<T::AccountId> where
Args: Encode,
Salt: AsRef<[u8]>,

Instantiates another contract.

Note

For more details visit: ink_env::instantiate_contract

pub fn restore_contract(
self,
account_id: T::AccountId,
code_hash: T::Hash,
rent_allowance: T::Balance,
filtered_keys: &[Key]
)

Restores a smart contract in tombstone state.

Note

For more details visit: ink_env::restore_contract

pub fn terminate_contract(self, beneficiary: T::AccountId) -> !

Terminates the existence of a smart contract.

Note

For more details visit: ink_env::terminate_contract

pub fn transfer(
self,
destination: T::AccountId,
value: T::Balance
) -> Result<()>

Transfers value from the contract to the destination account ID.

Note

For more details visit: ink_env::transfer

pub fn random(self, subject: &[u8]) -> (T::Hash, T::BlockNumber)

Returns a random hash seed.

Note

For more details visit: ink_env::random

pub fn hash_bytes<H>(self, input: &[u8]) -> <H as HashOutput>::Type where
H: CryptoHash,

Computes the hash of the given bytes using the cryptographic hash H.

Note

For more details visit: ink_env::hash_bytes

pub fn hash_encoded<H, V>(self, value: &V) -> <H as HashOutput>::Type where
H: CryptoHash,
V: Encode,

Computes the hash of the given SCALE encoded value using the cryptographic hash H.

Note

For more details visit: ink_env::hash_encoded